pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-rag"
version = "0.0.0"
description = "Local-first, transcript-centric Personal RAG"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "redis>=5.0",
  "rq>=1.16",
]

[tool.setuptools]
packages = ["app", "app.scripts"]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
]
